git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@1357 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Content/Effects/Global.inc.fx
@@ -1,8 +1,8 @@
 struct PSOutput
 {
-	half4 Color              : COLOR0;
-	half4 RenderChannelColor : COLOR1;
-	half4 DepthColor         : COLOR2;
+	float4 Color              : COLOR0;
+	float4 RenderChannelColor : COLOR1;
+	float4 DepthColor         : COLOR2;
 };
 
 const float DepthClipY = 600.0f;
