+++ src/ProjectMagma/ProjectMagma/Content/Effects/Sfx/ParticleSystem/Stateful/Implementations/IceSpike.fx
@@ -145,7 +145,7 @@ RenderParticlesVertexShaderOutput RenderIceSpikeVertexShader(
 
 
 //-----------------------------------------------------------------------------------------------------------
-half4 RenderIceSpikePixelShader(
+float4 RenderIceSpikePixelShader(
 	RenderParticlesVertexShaderOutput input,
 #ifdef XBOX
 	float2 particleCoordinate : SPRITETEXCOORD
