+++ src/ProjectMagma/ProjectMagma/Content/Effects/Sfx/ParticleSystem/Stateful/Implementations/Snow.fx
@@ -144,7 +144,7 @@ RenderSnowParticlesVertexShaderOutput RenderSnowVertexShader(
 
 
 //-----------------------------------------------------------------------------------------------------------
-half4 RenderSnowPixelShader(
+float4 RenderSnowPixelShader(
 	RenderSnowParticlesVertexShaderOutput input,
 #ifdef XBOX
 	float2 particleCoordinate : SPRITETEXCOORD
